Cleaned up and sorted javascript.
Only battlelog.bf3.inject.js is injected C# side to allow for more flexibility. Javascript will no longer be stored locally in production, instead hosted on a webserver to allow for up-to-the-minute updates. URL that serves Javascript will be configurable to allow future-proofing. For the forseeable future, javascript will be hosted on the gh-pages branch, however, Battlelogium will take any Javascript served to it from any server that hosts the files.